+++ src/main/java/de/cismet/cismap/custom/attributerule/SgSeeWrrlRuleSet.java
@@ -117,12 +117,12 @@ public class SgSeeWrrlRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Object getAdditionalFieldValue(final java.lang.String propertyName, final FeatureServiceFeature feature) {
-        Integer value = null;
+        Long value = null;
 
         final Geometry geom = ((Geometry)feature.getProperty("geom"));
 
         if (geom != null) {
-            value = (int)Math.round(geom.getArea());
+            value = Math.round(geom.getArea());
         }
 
         return value;
@@ -130,7 +130,7 @@ public class SgSeeWrrlRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Class getAdditionalFieldClass(final int index) {
-        return Integer.class;
+        return Long.class;
     }
 
     @Override
