+++ src/main/java/de/cismet/cismap/custom/attributerule/VwAlkMvRuleSet.java
@@ -114,12 +114,12 @@ public class VwAlkMvRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Object getAdditionalFieldValue(final java.lang.String propertyName, final FeatureServiceFeature feature) {
-        Integer value = null;
+        Long value = null;
 
         final Geometry geom = ((Geometry)feature.getProperty("geom"));
 
         if (geom != null) {
-            value = (int)Math.round(geom.getArea());
+            value = Math.round(geom.getArea());
         }
 
         return value;
@@ -127,7 +127,7 @@ public class VwAlkMvRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Class getAdditionalFieldClass(final int index) {
-        return Integer.class;
+        return Long.class;
     }
 
     @Override
