+++ src/main/java/de/cismet/cismap/custom/attributerule/VwDvgMvRuleSet.java
@@ -92,12 +92,12 @@ public class VwDvgMvRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Object getAdditionalFieldValue(final java.lang.String propertyName, final FeatureServiceFeature feature) {
-        Integer value = null;
+        Long value = null;
 
         final Geometry geom = ((Geometry)feature.getProperty("geom"));
 
         if (geom != null) {
-            value = (int)Math.round(geom.getArea());
+            value = Math.round(geom.getArea());
         }
 
         return value;
@@ -105,7 +105,7 @@ public class VwDvgMvRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Class getAdditionalFieldClass(final int index) {
-        return Integer.class;
+        return Long.class;
     }
 
     @Override
